@@ -1,6 +1,6 @@
 {
   "name": "algosigner-ui",
-  "version": "1.6.0",
+  "version": "1.6.1",
   "author": "https://developer.purestake.io",
   "repository": "https://github.com/PureStake/algosigner",
   "license": "MIT",
@@ -16,7 +16,7 @@
     "@fortawesome/fontawesome-free": "^5.15.0",
     "@ledgerhq/hw-app-algorand": "^5.51.1",
     "@ledgerhq/hw-transport-webhid": "^5.46.0",
-    "algosdk": "1.8.1",
+    "algosdk": "1.9.1",
     "history": "^5.0.0",
     "htm": "^3.0.4",
     "mobx": "^5.15.6",
@@ -45,7 +45,7 @@
     "sass-loader": "^8.0.2",
     "ts-jest": "^27.0.0",
     "ts-loader": "^7.0.1",
-    "typescript": "^3.9.7",
+    "typescript": "^4.3.2",
     "webpack": "^4.44.1",
     "webpack-cli": "^3.3.11",
     "webpack-dev-server": "^3.11.2"
