@@ -1,7 +1,8 @@
 'use strict';
-const algosdk = require("algosdk");
-import encryptionWrap from "./encryptionWrap";
-import createNewAccount from "./account/createAccount.js";
-import { TransactionType } from "@algosigner/common/types/transaction";
+
+(BigInt as any).prototype.toJSON = function () {
+  return this.toString();
+};
+
 import Background from './background';
 Background.start();
