@@ -7,7 +7,7 @@ import { BaseValidatedTxnWrap } from './baseValidatedTxnWrap';
 class AssetTransferTx implements IAssetTransferTx {
   type: string = undefined;
   assetIndex: number = undefined;
-  amount: number = 0;
+  amount: BigInt = BigInt(0);
   from: string = undefined;
   to: string = undefined;
   closeRemainderTo?: string = null;
